+++ src/components/board/WordToGuess.js
