+++ src/components/canvas/Eraser.js
