+++ src/components/canvas/Palette.js
