+++ src/components/modals/RoundOverModal.js
@@ -5,7 +5,11 @@ const RoundOverModal = (props) => {
 
   return (
     <div className={`round-over-modal ${showHideClassName}`}>
-      <h2>The word was: <span className="word-was">{props.word}</span></h2>
+      <h2>
+        The word was:
+        {' '}
+        <span className="word-was">{props.word}</span>
+      </h2>
     </div>
   );
 };
